datadog_agent_config

Trait DatadogConfigWitness

pub trait DatadogConfigWitness {

    // Required methods
    fn consume_additional_endpoints(
        &mut self,
        value: HashMap<String, Vec<String>>,
    );
    fn consume_agent_ipc_grpc_max_message_size(&mut self, value: i64);
    fn consume_aggregator_stop_timeout(&mut self, value: i64);
    fn consume_allow_arbitrary_tags(&mut self, value: bool);
    fn consume_api_key(&mut self, value: String);
    fn consume_apm_config_compute_stats_by_span_kind(&mut self, value: bool);
    fn consume_apm_config_enable_rare_sampler(&mut self, value: bool);
    fn consume_apm_config_error_tracking_standalone_enabled(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_errors_per_second(&mut self, value: f64);
    fn consume_apm_config_obfuscation_credit_cards_enabled(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_obfuscation_credit_cards_keep_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_credit_cards_luhn(&mut self, value: bool);
    fn consume_apm_config_obfuscation_elasticsearch_enabled(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_obfuscation_elasticsearch_keep_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_elasticsearch_obfuscate_sql_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_http_remove_paths_with_digits(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_obfuscation_http_remove_query_string(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_obfuscation_memcached_enabled(&mut self, value: bool);
    fn consume_apm_config_obfuscation_memcached_keep_command(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_obfuscation_mongodb_enabled(&mut self, value: bool);
    fn consume_apm_config_obfuscation_mongodb_keep_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_mongodb_obfuscate_sql_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_opensearch_enabled(&mut self, value: bool);
    fn consume_apm_config_obfuscation_opensearch_keep_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_opensearch_obfuscate_sql_values(
        &mut self,
        value: Vec<String>,
    );
    fn consume_apm_config_obfuscation_redis_enabled(&mut self, value: bool);
    fn consume_apm_config_obfuscation_redis_remove_all_args(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_obfuscation_valkey_enabled(&mut self, value: bool);
    fn consume_apm_config_obfuscation_valkey_remove_all_args(
        &mut self,
        value: bool,
    );
    fn consume_apm_config_peer_tags(&mut self, value: Vec<String>);
    fn consume_apm_config_peer_tags_aggregation(&mut self, value: bool);
    fn consume_apm_config_probabilistic_sampler_enabled(&mut self, value: bool);
    fn consume_apm_config_probabilistic_sampler_sampling_percentage(
        &mut self,
        value: f64,
    );
    fn consume_apm_config_target_traces_per_second(&mut self, value: f64);
    fn consume_autoscaling_failover_enabled(&mut self, value: bool);
    fn consume_autoscaling_failover_metrics(&mut self, value: Vec<String>);
    fn consume_bind_host(&mut self, value: String);
    fn consume_cluster_agent_auth_token(&mut self, value: String);
    fn consume_cluster_agent_enabled(&mut self, value: bool);
    fn consume_cluster_agent_kubernetes_service_name(&mut self, value: String);
    fn consume_cluster_agent_url(&mut self, value: String);
    fn consume_cmd_port(&mut self, value: i64);
    fn consume_cri_connection_timeout(&mut self, value: i64);
    fn consume_cri_query_timeout(&mut self, value: i64);
    fn consume_data_plane_api_listen_address(&mut self, value: String);
    fn consume_data_plane_dogstatsd_aggregator_tag_filter_cache_capacity(
        &mut self,
        value: i64,
    );
    fn consume_data_plane_dogstatsd_enabled(&mut self, value: bool);
    fn consume_data_plane_enabled(&mut self, value: bool);
    fn consume_data_plane_log_file(&mut self, value: String);
    fn consume_data_plane_otlp_enabled(&mut self, value: bool);
    fn consume_data_plane_otlp_proxy_enabled(&mut self, value: bool);
    fn consume_data_plane_otlp_proxy_logs_enabled(&mut self, value: bool);
    fn consume_data_plane_otlp_proxy_metrics_enabled(&mut self, value: bool);
    fn consume_data_plane_otlp_proxy_receiver_protocols_grpc_endpoint(
        &mut self,
        value: String,
    );
    fn consume_data_plane_otlp_proxy_traces_enabled(&mut self, value: bool);
    fn consume_data_plane_remote_agent_enabled(&mut self, value: bool);
    fn consume_data_plane_secure_api_listen_address(&mut self, value: String);
    fn consume_data_plane_use_new_config_stream_endpoint(&mut self, value: bool);
    fn consume_dd_url(&mut self, value: String);
    fn consume_disable_file_logging(&mut self, value: bool);
    fn consume_dogstatsd_buffer_size(&mut self, value: i64);
    fn consume_dogstatsd_capture_depth(&mut self, value: i64);
    fn consume_dogstatsd_capture_path(&mut self, value: String);
    fn consume_dogstatsd_context_expiry_seconds(&mut self, value: i64);
    fn consume_dogstatsd_disable_verbose_logs(&mut self, value: bool);
    fn consume_dogstatsd_entity_id_precedence(&mut self, value: bool);
    fn consume_dogstatsd_eol_required(&mut self, value: Vec<String>);
    fn consume_dogstatsd_expiry_seconds(&mut self, value: i64);
    fn consume_dogstatsd_flush_incomplete_buckets(&mut self, value: bool);
    fn consume_dogstatsd_log_file(&mut self, value: String);
    fn consume_dogstatsd_log_file_max_rolls(&mut self, value: i64);
    fn consume_dogstatsd_log_file_max_size(&mut self, value: String);
    fn consume_dogstatsd_logging_enabled(&mut self, value: bool);
    fn consume_dogstatsd_mapper_cache_size(&mut self, value: i64);
    fn consume_dogstatsd_mapper_profiles(&mut self, value: Vec<Value>);
    fn consume_dogstatsd_metrics_stats_enable(&mut self, value: bool);
    fn consume_dogstatsd_no_aggregation_pipeline(&mut self, value: bool);
    fn consume_dogstatsd_non_local_traffic(&mut self, value: bool);
    fn consume_dogstatsd_origin_detection(&mut self, value: bool);
    fn consume_dogstatsd_origin_detection_client(&mut self, value: bool);
    fn consume_dogstatsd_origin_optout_enabled(&mut self, value: bool);
    fn consume_dogstatsd_pipe_name(&mut self, value: String);
    fn consume_dogstatsd_port(&mut self, value: i64);
    fn consume_dogstatsd_so_rcvbuf(&mut self, value: i64);
    fn consume_dogstatsd_socket(&mut self, value: Option<String>);
    fn consume_dogstatsd_stream_log_too_big(&mut self, value: bool);
    fn consume_dogstatsd_stream_socket(&mut self, value: String);
    fn consume_dogstatsd_string_interner_size(&mut self, value: i64);
    fn consume_dogstatsd_tag_cardinality(&mut self, value: String);
    fn consume_dogstatsd_tags(&mut self, value: Vec<String>);
    fn consume_dogstatsd_windows_pipe_security_descriptor(
        &mut self,
        value: String,
    );
    fn consume_enable_payloads_events(&mut self, value: bool);
    fn consume_enable_payloads_series(&mut self, value: bool);
    fn consume_enable_payloads_service_checks(&mut self, value: bool);
    fn consume_enable_payloads_sketches(&mut self, value: bool);
    fn consume_env(&mut self, value: String);
    fn consume_expected_tags_duration(&mut self, value: Duration);
    fn consume_forwarder_apikey_validation_interval(&mut self, value: i64);
    fn consume_forwarder_backoff_base(&mut self, value: i64);
    fn consume_forwarder_backoff_factor(&mut self, value: i64);
    fn consume_forwarder_backoff_max(&mut self, value: i64);
    fn consume_forwarder_connection_reset_interval(&mut self, value: i64);
    fn consume_forwarder_flush_to_disk_mem_ratio(&mut self, value: f64);
    fn consume_forwarder_high_prio_buffer_size(&mut self, value: i64);
    fn consume_forwarder_http_protocol(&mut self, value: String);
    fn consume_forwarder_max_concurrent_requests(&mut self, value: i64);
    fn consume_forwarder_num_workers(&mut self, value: i64);
    fn consume_forwarder_outdated_file_in_days(&mut self, value: i64);
    fn consume_forwarder_recovery_interval(&mut self, value: i64);
    fn consume_forwarder_recovery_reset(&mut self, value: bool);
    fn consume_forwarder_retry_queue_capacity_time_interval_sec(
        &mut self,
        value: i64,
    );
    fn consume_forwarder_retry_queue_max_size(&mut self, value: i64);
    fn consume_forwarder_retry_queue_payloads_max_size(&mut self, value: i64);
    fn consume_forwarder_stop_timeout(&mut self, value: i64);
    fn consume_forwarder_storage_max_disk_ratio(&mut self, value: f64);
    fn consume_forwarder_storage_max_size_in_bytes(&mut self, value: i64);
    fn consume_forwarder_storage_path(&mut self, value: String);
    fn consume_forwarder_timeout(&mut self, value: i64);
    fn consume_histogram_aggregates(&mut self, value: Vec<String>);
    fn consume_histogram_copy_to_distribution(&mut self, value: bool);
    fn consume_histogram_copy_to_distribution_prefix(&mut self, value: String);
    fn consume_histogram_percentiles(&mut self, value: Vec<String>);
    fn consume_log_file_max_rolls(&mut self, value: i64);
    fn consume_log_file_max_size(&mut self, value: String);
    fn consume_log_format_json(&mut self, value: bool);
    fn consume_log_format_rfc3339(&mut self, value: bool);
    fn consume_log_level(&mut self, value: String);
    fn consume_log_payloads(&mut self, value: bool);
    fn consume_log_to_console(&mut self, value: bool);
    fn consume_log_to_syslog(&mut self, value: bool);
    fn consume_metric_filterlist(&mut self, value: Vec<String>);
    fn consume_metric_filterlist_match_prefix(&mut self, value: bool);
    fn consume_metric_tag_filterlist(&mut self, value: Vec<Value>);
    fn consume_min_tls_version(&mut self, value: String);
    fn consume_multi_region_failover_api_key(&mut self, value: String);
    fn consume_multi_region_failover_dd_url(&mut self, value: String);
    fn consume_multi_region_failover_enabled(&mut self, value: bool);
    fn consume_multi_region_failover_failover_metrics(&mut self, value: bool);
    fn consume_multi_region_failover_metric_allowlist(
        &mut self,
        value: Vec<String>,
    );
    fn consume_multi_region_failover_site(&mut self, value: String);
    fn consume_no_proxy_nonexact_match(&mut self, value: bool);
    fn consume_observability_pipelines_worker_metrics_enabled(
        &mut self,
        value: bool,
    );
    fn consume_observability_pipelines_worker_metrics_url(
        &mut self,
        value: String,
    );
    fn consume_observability_pipelines_worker_metrics_use_v3_api_series(
        &mut self,
        value: bool,
    );
    fn consume_origin_detection_unified(&mut self, value: bool);
    fn consume_otlp_config_logs_enabled(&mut self, value: bool);
    fn consume_otlp_config_metrics_enabled(&mut self, value: bool);
    fn consume_otlp_config_metrics_histograms_mode(&mut self, value: String);
    fn consume_otlp_config_metrics_histograms_send_aggregation_metrics(
        &mut self,
        value: bool,
    );
    fn consume_otlp_config_metrics_resource_attributes_as_tags(
        &mut self,
        value: bool,
    );
    fn consume_otlp_config_metrics_summaries_mode(&mut self, value: String);
    fn consume_otlp_config_metrics_sums_cumulative_monotonic_mode(
        &mut self,
        value: String,
    );
    fn consume_otlp_config_metrics_sums_initial_cumulative_monotonic_value(
        &mut self,
        value: String,
    );
    fn consume_otlp_config_metrics_tags(&mut self, value: String);
    fn consume_otlp_config_receiver_protocols_grpc_endpoint(
        &mut self,
        value: String,
    );
    fn consume_otlp_config_receiver_protocols_grpc_max_recv_msg_size_mib(
        &mut self,
        value: i64,
    );
    fn consume_otlp_config_receiver_protocols_grpc_transport(
        &mut self,
        value: String,
    );
    fn consume_otlp_config_receiver_protocols_http_endpoint(
        &mut self,
        value: String,
    );
    fn consume_otlp_config_traces_enabled(&mut self, value: bool);
    fn consume_otlp_config_traces_internal_port(&mut self, value: i64);
    fn consume_otlp_config_traces_probabilistic_sampler_sampling_percentage(
        &mut self,
        value: f64,
    );
    fn consume_provider_kind(&mut self, value: String);
    fn consume_proxy_http(&mut self, value: String);
    fn consume_proxy_https(&mut self, value: String);
    fn consume_proxy_no_proxy(&mut self, value: Vec<String>);
    fn consume_serializer_compressor_kind(&mut self, value: String);
    fn consume_serializer_experimental_use_v3_api_compression_level(
        &mut self,
        value: i64,
    );
    fn consume_serializer_experimental_use_v3_api_series_beta_route(
        &mut self,
        value: String,
    );
    fn consume_serializer_experimental_use_v3_api_series_endpoints(
        &mut self,
        value: Vec<String>,
    );
    fn consume_serializer_experimental_use_v3_api_series_shadow_sample_rate(
        &mut self,
        value: f64,
    );
    fn consume_serializer_experimental_use_v3_api_series_shadow_sites(
        &mut self,
        value: Vec<String>,
    );
    fn consume_serializer_experimental_use_v3_api_series_use_beta(
        &mut self,
        value: bool,
    );
    fn consume_serializer_experimental_use_v3_api_series_validate(
        &mut self,
        value: bool,
    );
    fn consume_serializer_experimental_use_v3_api_sketches_endpoints(
        &mut self,
        value: Vec<String>,
    );
    fn consume_serializer_experimental_use_v3_api_sketches_validate(
        &mut self,
        value: bool,
    );
    fn consume_serializer_max_payload_size(&mut self, value: i64);
    fn consume_serializer_max_series_payload_size(&mut self, value: i64);
    fn consume_serializer_max_series_points_per_payload(&mut self, value: i64);
    fn consume_serializer_max_series_uncompressed_payload_size(
        &mut self,
        value: i64,
    );
    fn consume_serializer_max_uncompressed_payload_size(&mut self, value: i64);
    fn consume_serializer_zstd_compressor_level(&mut self, value: i64);
    fn consume_site(&mut self, value: String);
    fn consume_skip_ssl_validation(&mut self, value: bool);
    fn consume_sslkeylogfile(&mut self, value: String);
    fn consume_statsd_forward_host(&mut self, value: String);
    fn consume_statsd_forward_port(&mut self, value: i64);
    fn consume_statsd_metric_blocklist(&mut self, value: Vec<String>);
    fn consume_statsd_metric_blocklist_match_prefix(&mut self, value: bool);
    fn consume_statsd_metric_namespace(&mut self, value: String);
    fn consume_statsd_metric_namespace_blacklist(&mut self, value: Vec<String>);
    fn consume_syslog_rfc(&mut self, value: bool);
    fn consume_syslog_uri(&mut self, value: String);
    fn consume_telemetry_dogstatsd_origin(&mut self, value: bool);
    fn consume_use_proxy_for_cloud_metadata(&mut self, value: bool);
    fn consume_use_v2_api_series(&mut self, value: bool);
    fn consume_use_v3_api_series_enabled(&mut self, value: String);
    fn consume_use_v3_api_series_endpoints(&mut self, value: Map<String, Value>);
    fn consume_vector_metrics_enabled(&mut self, value: bool);
    fn consume_vector_metrics_url(&mut self, value: String);
    fn consume_vector_metrics_use_v3_api_series(&mut self, value: bool);
    fn consume_vsock_addr(&mut self, value: String);

    // Provided method
    fn translate_errors(&mut self) -> Vec<TranslateError> { ... }

}

Witness trait over every supported Datadog configuration key.

One consume_<key> method per support: full / support: partial key from the Datadog schema overlay. This trait must be implemented by hand to ensure that each supported value from Datadog configuration finds a home in Saluki configuration.

Required Methods

fn consume_additional_endpoints(&mut self, value: HashMap<String, Vec<String>>)

fn consume_agent_ipc_grpc_max_message_size(&mut self, value: i64)

fn consume_aggregator_stop_timeout(&mut self, value: i64)

fn consume_allow_arbitrary_tags(&mut self, value: bool)

fn consume_api_key(&mut self, value: String)

fn consume_apm_config_compute_stats_by_span_kind(&mut self, value: bool)

fn consume_apm_config_enable_rare_sampler(&mut self, value: bool)

fn consume_apm_config_error_tracking_standalone_enabled(&mut self, value: bool)

fn consume_apm_config_errors_per_second(&mut self, value: f64)

fn consume_apm_config_obfuscation_credit_cards_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_credit_cards_keep_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_credit_cards_luhn(&mut self, value: bool)

fn consume_apm_config_obfuscation_elasticsearch_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_elasticsearch_keep_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_elasticsearch_obfuscate_sql_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_http_remove_paths_with_digits( &mut self, value: bool, )

fn consume_apm_config_obfuscation_http_remove_query_string( &mut self, value: bool, )

fn consume_apm_config_obfuscation_memcached_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_memcached_keep_command(&mut self, value: bool)

fn consume_apm_config_obfuscation_mongodb_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_mongodb_keep_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_mongodb_obfuscate_sql_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_opensearch_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_opensearch_keep_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_opensearch_obfuscate_sql_values( &mut self, value: Vec<String>, )

fn consume_apm_config_obfuscation_redis_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_redis_remove_all_args(&mut self, value: bool)

fn consume_apm_config_obfuscation_valkey_enabled(&mut self, value: bool)

fn consume_apm_config_obfuscation_valkey_remove_all_args(&mut self, value: bool)

fn consume_apm_config_peer_tags(&mut self, value: Vec<String>)

fn consume_apm_config_peer_tags_aggregation(&mut self, value: bool)

fn consume_apm_config_probabilistic_sampler_enabled(&mut self, value: bool)

fn consume_apm_config_probabilistic_sampler_sampling_percentage( &mut self, value: f64, )

fn consume_apm_config_target_traces_per_second(&mut self, value: f64)

fn consume_autoscaling_failover_enabled(&mut self, value: bool)

fn consume_autoscaling_failover_metrics(&mut self, value: Vec<String>)

fn consume_bind_host(&mut self, value: String)

fn consume_cluster_agent_auth_token(&mut self, value: String)

fn consume_cluster_agent_enabled(&mut self, value: bool)

fn consume_cluster_agent_kubernetes_service_name(&mut self, value: String)

fn consume_cluster_agent_url(&mut self, value: String)

fn consume_cmd_port(&mut self, value: i64)

fn consume_cri_connection_timeout(&mut self, value: i64)

fn consume_cri_query_timeout(&mut self, value: i64)

fn consume_data_plane_api_listen_address(&mut self, value: String)

fn consume_data_plane_dogstatsd_aggregator_tag_filter_cache_capacity( &mut self, value: i64, )

fn consume_data_plane_dogstatsd_enabled(&mut self, value: bool)

fn consume_data_plane_enabled(&mut self, value: bool)

fn consume_data_plane_log_file(&mut self, value: String)

fn consume_data_plane_otlp_enabled(&mut self, value: bool)

fn consume_data_plane_otlp_proxy_enabled(&mut self, value: bool)

fn consume_data_plane_otlp_proxy_logs_enabled(&mut self, value: bool)

fn consume_data_plane_otlp_proxy_metrics_enabled(&mut self, value: bool)

fn consume_data_plane_otlp_proxy_receiver_protocols_grpc_endpoint( &mut self, value: String, )

fn consume_data_plane_otlp_proxy_traces_enabled(&mut self, value: bool)

fn consume_data_plane_remote_agent_enabled(&mut self, value: bool)

fn consume_data_plane_secure_api_listen_address(&mut self, value: String)

fn consume_data_plane_use_new_config_stream_endpoint(&mut self, value: bool)

fn consume_dd_url(&mut self, value: String)

fn consume_disable_file_logging(&mut self, value: bool)

fn consume_dogstatsd_buffer_size(&mut self, value: i64)

fn consume_dogstatsd_capture_depth(&mut self, value: i64)

fn consume_dogstatsd_capture_path(&mut self, value: String)

fn consume_dogstatsd_context_expiry_seconds(&mut self, value: i64)

fn consume_dogstatsd_disable_verbose_logs(&mut self, value: bool)

fn consume_dogstatsd_entity_id_precedence(&mut self, value: bool)

fn consume_dogstatsd_eol_required(&mut self, value: Vec<String>)

fn consume_dogstatsd_expiry_seconds(&mut self, value: i64)

fn consume_dogstatsd_flush_incomplete_buckets(&mut self, value: bool)

fn consume_dogstatsd_log_file(&mut self, value: String)

fn consume_dogstatsd_log_file_max_rolls(&mut self, value: i64)

fn consume_dogstatsd_log_file_max_size(&mut self, value: String)

fn consume_dogstatsd_logging_enabled(&mut self, value: bool)

fn consume_dogstatsd_mapper_cache_size(&mut self, value: i64)

fn consume_dogstatsd_mapper_profiles(&mut self, value: Vec<Value>)

fn consume_dogstatsd_metrics_stats_enable(&mut self, value: bool)

fn consume_dogstatsd_no_aggregation_pipeline(&mut self, value: bool)

fn consume_dogstatsd_non_local_traffic(&mut self, value: bool)

fn consume_dogstatsd_origin_detection(&mut self, value: bool)

fn consume_dogstatsd_origin_detection_client(&mut self, value: bool)

fn consume_dogstatsd_origin_optout_enabled(&mut self, value: bool)

fn consume_dogstatsd_pipe_name(&mut self, value: String)

fn consume_dogstatsd_port(&mut self, value: i64)

fn consume_dogstatsd_so_rcvbuf(&mut self, value: i64)

fn consume_dogstatsd_socket(&mut self, value: Option<String>)

fn consume_dogstatsd_stream_log_too_big(&mut self, value: bool)

fn consume_dogstatsd_stream_socket(&mut self, value: String)

fn consume_dogstatsd_string_interner_size(&mut self, value: i64)

fn consume_dogstatsd_tag_cardinality(&mut self, value: String)

fn consume_dogstatsd_tags(&mut self, value: Vec<String>)

fn consume_dogstatsd_windows_pipe_security_descriptor(&mut self, value: String)

fn consume_enable_payloads_events(&mut self, value: bool)

fn consume_enable_payloads_series(&mut self, value: bool)

fn consume_enable_payloads_service_checks(&mut self, value: bool)

fn consume_enable_payloads_sketches(&mut self, value: bool)

fn consume_env(&mut self, value: String)

fn consume_expected_tags_duration(&mut self, value: Duration)

fn consume_forwarder_apikey_validation_interval(&mut self, value: i64)

fn consume_forwarder_backoff_base(&mut self, value: i64)

fn consume_forwarder_backoff_factor(&mut self, value: i64)

fn consume_forwarder_backoff_max(&mut self, value: i64)

fn consume_forwarder_connection_reset_interval(&mut self, value: i64)

fn consume_forwarder_flush_to_disk_mem_ratio(&mut self, value: f64)

fn consume_forwarder_high_prio_buffer_size(&mut self, value: i64)

fn consume_forwarder_http_protocol(&mut self, value: String)

fn consume_forwarder_max_concurrent_requests(&mut self, value: i64)

fn consume_forwarder_num_workers(&mut self, value: i64)

fn consume_forwarder_outdated_file_in_days(&mut self, value: i64)

fn consume_forwarder_recovery_interval(&mut self, value: i64)

fn consume_forwarder_recovery_reset(&mut self, value: bool)

fn consume_forwarder_retry_queue_capacity_time_interval_sec( &mut self, value: i64, )

fn consume_forwarder_retry_queue_max_size(&mut self, value: i64)

fn consume_forwarder_retry_queue_payloads_max_size(&mut self, value: i64)

fn consume_forwarder_stop_timeout(&mut self, value: i64)

fn consume_forwarder_storage_max_disk_ratio(&mut self, value: f64)

fn consume_forwarder_storage_max_size_in_bytes(&mut self, value: i64)

fn consume_forwarder_storage_path(&mut self, value: String)

fn consume_forwarder_timeout(&mut self, value: i64)

fn consume_histogram_aggregates(&mut self, value: Vec<String>)

fn consume_histogram_copy_to_distribution(&mut self, value: bool)

fn consume_histogram_copy_to_distribution_prefix(&mut self, value: String)

fn consume_histogram_percentiles(&mut self, value: Vec<String>)

fn consume_log_file_max_rolls(&mut self, value: i64)

fn consume_log_file_max_size(&mut self, value: String)

fn consume_log_format_json(&mut self, value: bool)

fn consume_log_format_rfc3339(&mut self, value: bool)

fn consume_log_level(&mut self, value: String)

fn consume_log_payloads(&mut self, value: bool)

fn consume_log_to_console(&mut self, value: bool)

fn consume_log_to_syslog(&mut self, value: bool)

fn consume_metric_filterlist(&mut self, value: Vec<String>)

fn consume_metric_filterlist_match_prefix(&mut self, value: bool)

fn consume_metric_tag_filterlist(&mut self, value: Vec<Value>)

fn consume_min_tls_version(&mut self, value: String)

fn consume_multi_region_failover_api_key(&mut self, value: String)

fn consume_multi_region_failover_dd_url(&mut self, value: String)

fn consume_multi_region_failover_enabled(&mut self, value: bool)

fn consume_multi_region_failover_failover_metrics(&mut self, value: bool)

fn consume_multi_region_failover_metric_allowlist(&mut self, value: Vec<String>)

fn consume_multi_region_failover_site(&mut self, value: String)

fn consume_no_proxy_nonexact_match(&mut self, value: bool)

fn consume_observability_pipelines_worker_metrics_enabled( &mut self, value: bool, )

fn consume_observability_pipelines_worker_metrics_url(&mut self, value: String)

fn consume_observability_pipelines_worker_metrics_use_v3_api_series( &mut self, value: bool, )

fn consume_origin_detection_unified(&mut self, value: bool)

fn consume_otlp_config_logs_enabled(&mut self, value: bool)

fn consume_otlp_config_metrics_enabled(&mut self, value: bool)

fn consume_otlp_config_metrics_histograms_mode(&mut self, value: String)

fn consume_otlp_config_metrics_histograms_send_aggregation_metrics( &mut self, value: bool, )

fn consume_otlp_config_metrics_resource_attributes_as_tags( &mut self, value: bool, )

fn consume_otlp_config_metrics_summaries_mode(&mut self, value: String)

fn consume_otlp_config_metrics_sums_cumulative_monotonic_mode( &mut self, value: String, )

fn consume_otlp_config_metrics_sums_initial_cumulative_monotonic_value( &mut self, value: String, )

fn consume_otlp_config_metrics_tags(&mut self, value: String)

fn consume_otlp_config_receiver_protocols_grpc_endpoint( &mut self, value: String, )

fn consume_otlp_config_receiver_protocols_grpc_max_recv_msg_size_mib( &mut self, value: i64, )

fn consume_otlp_config_receiver_protocols_grpc_transport( &mut self, value: String, )

fn consume_otlp_config_receiver_protocols_http_endpoint( &mut self, value: String, )

fn consume_otlp_config_traces_enabled(&mut self, value: bool)

fn consume_otlp_config_traces_internal_port(&mut self, value: i64)

fn consume_otlp_config_traces_probabilistic_sampler_sampling_percentage( &mut self, value: f64, )

fn consume_provider_kind(&mut self, value: String)

fn consume_proxy_http(&mut self, value: String)

fn consume_proxy_https(&mut self, value: String)

fn consume_proxy_no_proxy(&mut self, value: Vec<String>)

fn consume_serializer_compressor_kind(&mut self, value: String)

fn consume_serializer_experimental_use_v3_api_compression_level( &mut self, value: i64, )

fn consume_serializer_experimental_use_v3_api_series_beta_route( &mut self, value: String, )

fn consume_serializer_experimental_use_v3_api_series_endpoints( &mut self, value: Vec<String>, )

fn consume_serializer_experimental_use_v3_api_series_shadow_sample_rate( &mut self, value: f64, )

fn consume_serializer_experimental_use_v3_api_series_shadow_sites( &mut self, value: Vec<String>, )

fn consume_serializer_experimental_use_v3_api_series_use_beta( &mut self, value: bool, )

fn consume_serializer_experimental_use_v3_api_series_validate( &mut self, value: bool, )

fn consume_serializer_experimental_use_v3_api_sketches_endpoints( &mut self, value: Vec<String>, )

fn consume_serializer_experimental_use_v3_api_sketches_validate( &mut self, value: bool, )

fn consume_serializer_max_payload_size(&mut self, value: i64)

fn consume_serializer_max_series_payload_size(&mut self, value: i64)

fn consume_serializer_max_series_points_per_payload(&mut self, value: i64)

fn consume_serializer_max_series_uncompressed_payload_size( &mut self, value: i64, )

fn consume_serializer_max_uncompressed_payload_size(&mut self, value: i64)

fn consume_serializer_zstd_compressor_level(&mut self, value: i64)

fn consume_site(&mut self, value: String)

fn consume_skip_ssl_validation(&mut self, value: bool)

fn consume_sslkeylogfile(&mut self, value: String)

fn consume_statsd_forward_host(&mut self, value: String)

fn consume_statsd_forward_port(&mut self, value: i64)

fn consume_statsd_metric_blocklist(&mut self, value: Vec<String>)

fn consume_statsd_metric_blocklist_match_prefix(&mut self, value: bool)

fn consume_statsd_metric_namespace(&mut self, value: String)

fn consume_statsd_metric_namespace_blacklist(&mut self, value: Vec<String>)

fn consume_syslog_rfc(&mut self, value: bool)

fn consume_syslog_uri(&mut self, value: String)

fn consume_telemetry_dogstatsd_origin(&mut self, value: bool)

fn consume_use_proxy_for_cloud_metadata(&mut self, value: bool)

fn consume_use_v2_api_series(&mut self, value: bool)

fn consume_use_v3_api_series_enabled(&mut self, value: String)

fn consume_use_v3_api_series_endpoints(&mut self, value: Map<String, Value>)

fn consume_vector_metrics_enabled(&mut self, value: bool)

fn consume_vector_metrics_url(&mut self, value: String)

fn consume_vector_metrics_use_v3_api_series(&mut self, value: bool)

fn consume_vsock_addr(&mut self, value: String)

Provided Methods

fn translate_errors(&mut self) -> Vec<TranslateError>

Drains every translation error recorded while consuming.